+++ src/lib/validate.h
@@ -29,6 +29,7 @@ bool vcb_ispathfilename(sds data);
 bool vcb_isuri(sds data);
 bool vcb_isstreamuri(sds data);
 bool vcb_iscolumn(sds data);
+bool vcb_iscompareop(sds data);
 bool vcb_istaglist(sds data);
 bool vcb_ismpdtag(sds data);
 bool vcb_ismpdtag_or_any(sds data);